==> R/03_simulate_NS.qmd <==
---
title: "Simulating NanoString nCoutner data"
author: "Puck Quarles van Ufford"
format: html
---

## Packages

```{r}
#| message: false
#| label: libraries

library("tidyverse")
library("ggVennDiagram")
```

## Load data

```{r}
#| message: false
#| label: load_data

TCGA_exprs <- read_delim("../data/02_TCGA_exprs_clean.tsv")

NS_genes <- read_delim("../data/01_NS_genes.tsv", delim = "\t")
```

## Evaluate the coverage of NanoString genes in TCGA data

```{r}
#| label: eval_coverage
#| message: false

ggVennDiagram(list(
  Genes_TCGA = TCGA_exprs$feature,
  Genes_NS = NS_genes$ID
)) + 
  coord_flip() + 
  theme(legend.position = "None") + 
  labs(
    title = "Coverage of NanoString genes in TCGA dataset"
  )
```

Not every gene in the list has an exact match in the TCGA expression dataset, but the coverage is adequate for the purposes of this project.

## Subset the expression data 

```{r}
#| label: subset_exprs

TCGA_NS_subset <- TCGA_exprs |> 
  subset(feature %in% NS_genes$ID)
```

## Save data

```{r}
#| label: save_data

write_tsv(TCGA_NS_subset,
          file = "../data/03_TCGA_NS_sim.tsv")
```
