==> R/06_rDTC.qmd <==
---
title: "Classification with a ranked distance-to-centroid model"
author: "Puck Quarles van Ufford"
format: html
---

## Packages

```{r}
#| message: false
#| label: libraries

library("tidyverse")
library("rankdist")
library("caret")
```

## Load data

```{r}
#| message: false
#| label: load_data

TCGA_exprs <- read_delim("../data/02_TCGA_exprs_clean.tsv")

NS_exprs <- read_delim("../data/03_TCGA_NS_sim.tsv")

cms_labels <- read_delim("../data/02_cms_labels_clean.tsv")
```

Reformat the expression data:

```{r}
#| label: reformat_data

TCGA_exprs <- TCGA_exprs |> 
  column_to_rownames("feature")

NS_exprs <- NS_exprs |> 
  column_to_rownames("feature")
```

## Classification using ranked distance-to-centroid with leave-one-out cross validation

### TCGA data

```{r}
#| label: rDTC_TCGA

pred_table <- data.frame(matrix(ncol = 3,
                                nrow = 0))

for(i in 1:ncol(TCGA_exprs)) {
  train_exprs <- TCGA_exprs[-i]
  train_CMS <- cms_labels[-i,]$cms
  test_exprs <- TCGA_exprs[i]
  test_CMS <- cms_labels[i,]$cms
  
  centroids <- NULL
  for (class in unique(train_CMS)) {
    class_centroid <- rowMeans(train_exprs[,train_CMS==class])
    centroids <- cbind(centroids, class_centroid)
  }
  colnames(centroids) <- unique(train_CMS)
  centroids_ranked <- apply(centroids, 2, rank)
  test_exprs_ranked <- apply(test_exprs, 2, rank)
  d <- t(apply(X = test_exprs_ranked, 
               MARGIN = 2, 
               FUN = function(x) DistanceBlock(
                 mat = t(centroids_ranked), 
                 r = x
                 )))
  colnames(d) <- colnames(centroids_ranked)
  CMS_pred <- apply(d, 1, which.min)
  CMS_pred <- colnames(d)[CMS_pred]
  pred_table <- rbind(pred_table, c(test_CMS, CMS_pred, test_CMS==CMS_pred))
}

colnames(pred_table) <- c("CMS_true", "CMS_pred", "pred_correct")

table(pred_table$pred_correct)

NOLBL_count <- pred_table |> 
  filter(CMS_pred == "NOLBL") |> 
  nrow()

classification_rate_tcga <- c(nrow(pred_table), nrow(pred_table) - NOLBL_count, NOLBL_count)

confusion_tcga <- confusionMatrix(data = factor(pred_table$CMS_pred), 
                                   reference = factor(pred_table$CMS_true), 
                                   mode = "everything")

confusion_tcga
overall_res_tcga <- confusion_tcga$overall
by_class_res_tcga <- confusion_tcga$byClass
```

### NanoString data

```{r}
#| label: DTC_NS

pred_table <- data.frame(matrix(ncol = 3,
                                nrow = 0))

for(i in 1:ncol(NS_exprs)) {
  train_exprs <- NS_exprs[-i]
  train_CMS <- cms_labels[-i,]$cms
  test_exprs <- NS_exprs[i]
  test_CMS <- cms_labels[i,]$cms
  
  centroids <- NULL
  for (class in unique(train_CMS)) {
    class_centroid <- rowMeans(train_exprs[,train_CMS==class])
    centroids <- cbind(centroids, class_centroid)
  }
  colnames(centroids) <- unique(train_CMS)
  centroids_ranked <- apply(centroids, 2, rank)
  test_exprs_ranked <- apply(test_exprs, 2, rank)
  d <- t(apply(X = test_exprs_ranked, 
               MARGIN = 2, 
               FUN = function(x) DistanceBlock(
                 mat = t(centroids_ranked), 
                 r = x
                 )))
  colnames(d) <- colnames(centroids_ranked)
  CMS_pred <- apply(d, 1, which.min)
  CMS_pred <- colnames(d)[CMS_pred]
  pred_table <- rbind(pred_table, c(test_CMS, CMS_pred, test_CMS==CMS_pred))
}

colnames(pred_table) <- c("CMS_true", "CMS_pred", "pred_correct")

table(pred_table$pred_correct)

NOLBL_count <- pred_table |> 
  filter(CMS_pred == "NOLBL") |> 
  nrow()

classification_rate_ns <- c(nrow(pred_table), nrow(pred_table) - NOLBL_count, NOLBL_count)

confusion_ns <- confusionMatrix(data = factor(pred_table$CMS_pred),
                                reference = factor(pred_table$CMS_true), 
                                mode = "everything")

confusion_ns
overall_res_ns <- confusion_ns$overall
by_class_res_ns<- confusion_ns$byClass
```

## Save results

```{r}
overall_res <- overall_res_tcga |> 
  rbind(overall_res_ns) |> 
  as_tibble() |> 
  mutate(data = c("TCGA", "NS"),
         model = "rDTC")

write_csv(overall_res, file = "../results/06.rDTC_overall.csv")

by_class_res <- by_class_res_tcga |> 
  rbind(by_class_res_ns) |> 
  as_tibble()|> 
  mutate(class = c("CMS1", "CMS2", "CMS3", "CMS4", "NOLBL", "CMS1", "CMS2", "CMS3", "CMS4", "NOLBL"),
         data = c("TCGA", "TCGA", "TCGA", "TCGA", "TCGA", "NS", "NS", "NS", "NS", "NS"),
         model = "rDTC")

write_csv(by_class_res, file = "../results/06.rDTC_byclass.csv")

classification_rate <- classification_rate_tcga |> 
  rbind(classification_rate_ns) |> 
  as_tibble() |> 
  mutate(data = c("TCGA", "NS"),
         model = "rDTC")
colnames(classification_rate) <- c("n_samples", "n_classified", "n_NOLBL", "data", "model")

write_csv(classification_rate, file = "../results/06.rDTC_classification_rate.csv")
```
